arrays are flying every which way!
words and things. hi people. other words. i need new music to program
to. no, but if someone is reading this, listen closely.... okay so...
still listening?? good. it's nice that your listening, because I can
load the array from parse. it prints to console. but when I try to use
it to build my graph,  things mess up. you're a good listener. sleep
time. for sammy